factor out getModeFromExtension and related code to shared/
