+++ web/src/repo/compare/FileDiffConnection.tsx
@@ -1,9 +1,9 @@
 import * as React from 'react'
 import { TextDocumentItem } from '../../../../shared/src/api/client/types/textDocument'
 import * as GQL from '../../../../shared/src/graphqlschema'
+import { getModeFromPath } from '../../../../shared/src/languages'
 import { Connection, FilteredConnection } from '../../components/FilteredConnection'
 import { ExtensionsDocumentsProps } from '../../extensions/environment/ExtensionsEnvironment'
-import { getModeFromPath } from '../../util'
 import { ErrorLike, isErrorLike } from '../../util/errors'
 import { FileDiffNodeProps } from './FileDiffNode'
