+++ web/src/repo/routes.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Redirect, RouteComponentProps } from 'react-router'
+import { getModeFromPath } from '../../../shared/src/languages'
 import { ResizablePanel } from '../panel/Panel'
-import { getModeFromPath } from '../util'
 import { formatHash, isLegacyFragment, parseHash } from '../util/url'
 import { BlobPage } from './blob/BlobPage'
 import { RepositoryCommitsPage } from './commits/RepositoryCommitsPage'
