+++ web/src/search/input/CodeIntellifyBlob.tsx
@@ -14,12 +14,12 @@ import { Subject, Subscription } from 'rxjs'
 import { catchError, filter, map, withLatestFrom } from 'rxjs/operators'
 import { Position } from 'vscode-languageserver-types'
 import * as GQL from '../../../../shared/src/graphqlschema'
+import { getModeFromPath } from '../../../../shared/src/languages'
 import { getHover, getJumpURL } from '../../backend/features'
 import { LSPTextDocumentPositionParams } from '../../backend/lsp'
 import { ExtensionsDocumentsProps } from '../../extensions/environment/ExtensionsEnvironment'
 import { ExtensionsControllerProps } from '../../extensions/ExtensionsClientCommonContext'
 import { fetchBlob } from '../../repo/blob/BlobPage'
-import { getModeFromPath } from '../../util'
 import { ErrorLike, isErrorLike } from '../../util/errors'
 import { isDefined, propertyIsDefined } from '../../util/types'
