+++ web/src/tracking/eventLogger.tsx
@@ -1,10 +1,10 @@
 import { matchPath } from 'react-router'
 import uuid from 'uuid'
 import * as GQL from '../../../shared/src/graphqlschema'
+import { getPathExtension } from '../../../shared/src/languages'
 import { authenticatedUser } from '../auth'
 import { parseBrowserRepoURL } from '../repo'
 import { repoRevRoute } from '../routes'
-import { getPathExtension } from '../util'
 import { browserExtensionMessageReceived, handleQueryEvents, pageViewQueryParameters } from './analyticsUtils'
 import { serverAdmin } from './services/serverAdminWrapper'
 import { telligent } from './services/telligentWrapper'
